+++ Runtime/DataStructures/ItemLocalizedData.cs
@@ -1,3 +1,4 @@
+using System;
 using TriInspector;
 using UnityEngine;
 using UnityEngine.Localization;
@@ -7,13 +8,30 @@ namespace GameUtils
     [DeclareBoxGroup("localization", Title = "Localization")]
     public abstract class ItemLocalizedData : ItemTagsData
     {
-        [SerializeField, Group("internal")] private string _internalItemName = null;
+        [SerializeField, Group("internal")] private string _internalItemName = "";
         [SerializeField, Group("localization")] private LocalizedString _itemName = null;
         [SerializeField, Group("localization")] private LocalizedString _itemDescription = null;
 
         //
+        private string SafeLocalizedString(LocalizedString localizedString)
+        {
+            string value;
+            try
+            {
+                value = localizedString.GetLocalizedString();
+            }
+            catch (Exception)
+            {
+                return "NO_LOCALIZED_STRING";
+            }
+
+            //
+            return value;
+        }
+
+        //
         public string InternalName => _internalItemName;
-        public string Name => _itemName.GetLocalizedString();
-        public string Description => _itemDescription.GetLocalizedString();
+        public string Name => SafeLocalizedString(_itemName);
+        public string Description => SafeLocalizedString(_itemDescription);
     }
 }